+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- [fj-doc-base-kotlin] fj-script-helper version set to 2.0.0
+
 ## [8.10.0] - 2024-10-23
 
 ### Added